third_party/WebKit/LayoutTests/platform/win/external/wpt/XMLHttpRequest/setrequestheader-content-type-expected.txt

This is a testharness.js-based test.
FAIL setRequestHeader("") sends a blank string assert_equals: expected "Content-Type: \n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL setRequestHeader(" ") sends the string " " assert_equals: expected "Content-Type: \n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL setRequestHeader(null) sends the string "null" assert_equals: expected "Content-Type: null\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL setRequestHeader(undefined) sends the string "undefined" assert_equals: expected "Content-Type: undefined\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL String request has correct default Content-Type of "text/plain;charset=UTF-8" assert_equals: expected "Content-Type: text/plain;charset=UTF-8\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL String request keeps setRequestHeader() Content-Type, with charset adjusted to UTF-8 assert_equals: expected "Content-Type: text/javascript;charset=UTF-8\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL XML Document request respects setRequestHeader("") assert_equals: expected "Content-Type: \n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL XML Document request has correct default Content-Type of "application/xml;charset=UTF-8" assert_equals: expected "Content-Type: application/xml;charset=UTF-8\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL XML Document request keeps setRequestHeader() Content-Type, with charset adjusted to UTF-8 assert_equals: expected "Content-Type: application/xhtml+xml;charset=UTF-8\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL HTML Document request respects setRequestHeader("") assert_equals: expected "Content-Type: \n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL HTML Document request has correct default Content-Type of "text/html;charset=UTF-8" assert_equals: expected "Content-Type: text/html;charset=UTF-8\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL HTML Document request keeps setRequestHeader() Content-Type, with charset adjusted to UTF-8 assert_equals: expected "Content-Type: text/html+junk;charset=UTF-8\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL Blob request respects setRequestHeader("") to be specified assert_equals: expected "Content-Type: \n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL Blob request with unset type sends no Content-Type without setRequestHeader() call assert_equals: expected "" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL Blob request with unset type keeps setRequestHeader() Content-Type and charset assert_equals: expected "Content-Type: application/xml;charset=ASCII\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL Blob request with set type uses that it for Content-Type unless setRequestHeader() assert_equals: expected "Content-Type: application/xml;charset=ascii\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL Blob request with set type keeps setRequestHeader() Content-Type and charset assert_equals: expected "Content-Type: application/xml+junk;charset=ASCII\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL ArrayBuffer request respects setRequestHeader("") assert_equals: expected "Content-Type: \n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL ArrayBuffer request sends no Content-Type without setRequestHeader() call assert_equals: expected "" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL ArrayBuffer request keeps setRequestHeader() Content-Type and charset assert_equals: expected "Content-Type: application/xml;charset=ASCII\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL ArrayBufferView request respects setRequestHeader("") assert_equals: expected "Content-Type: \n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL ArrayBufferView request sends no Content-Type without setRequestHeader() call assert_equals: expected "" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL ArrayBufferView request keeps setRequestHeader() Content-Type and charset assert_equals: expected "Content-Type: application/xml;charset=ASCII\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL FormData request respects setRequestHeader("") assert_equals: expected "Content-Type: \n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL FormData request has correct default Content-Type of "multipart/form-data; boundary=_" assert_regexp_match: expected object "/multipart\/form-data; boundary=(.*)/" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL FormData request keeps setRequestHeader() Content-Type and charset assert_equals: expected "Content-Type: application/xml;charset=ASCII\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL URLSearchParams respects setRequestHeader("") assert_equals: expected "Content-Type: \n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL URLSearchParams request has correct default Content-Type of "application/x-www-form-urlencoded;charset=UTF-8" assert_equals: expected "Content-Type: application/x-www-form-urlencoded;charset=UTF-8\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL URLSearchParams request keeps setRequestHeader() Content-Type, with charset adjusted to UTF-8 assert_equals: expected "Content-Type: application/xml;charset=UTF-8\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL ReadableStream request respects setRequestHeader("") assert_equals: expected "Content-Type: \n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL ReadableStream request with under type sends no Content-Type without setRequestHeader() call assert_equals: expected "" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL ReadableStream request keeps setRequestHeader() Content-Type and charset assert_equals: expected "Content-Type: application/xml;charset=ASCII\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
Harness: the test ran to completion.
